[backend] refactor route for routes
