--- starter-project-backend-g32/src/routes/routes.ts
@@ -1,7 +1,0 @@
-import express from 'express';
-import commentRouter from './comment.routes';
-const router = express.Router();
-
-router.use("/comment", commentRouter);
-
-module.exports = router;
